Code_Competence_02/praktikum/requirements.txt
pandas==1.3.3
apache-airflow==2.2.3
pyarrow==5.0.0
google-cloud-storage